Støtte for frittstående PageNavigationMenu part. Refactor av navigasjonsmenyen for gjenbruk mellom part og produktside-komponent
