+++ src/components/layouts/page-with-side-menus/left-menu-section/LeftMenuSection.tsx
@@ -3,8 +3,8 @@ import { BEM, classNames } from '../../../../utils/classnames';
 import Region from '../../Region';
 import { RegionProps } from '../../../../types/component-props/layouts';
 import { ContentProps } from '../../../../types/content-props/_content-common';
-import { AccordionAnchorMenu, AccordionAnchorMenuItem } from '@navikt/ds-react';
-import { AnchorLink } from '../../../../types/component-props/pages/page-with-side-menus';
+import { PageNavigationMenu } from '../../../_common/page-navigation-menu/PageNavigationMenu';
+import { AnchorLink } from '../../../../types/component-props/parts/page-navigation-menu';
 import './LeftMenuSection.less';
 
 const bem = BEM('left-menu');
@@ -24,13 +24,10 @@ export const LeftMenuSection = ({
 }: Props) => {
     return (
         <div className={classNames(bem())}>
-            <AccordionAnchorMenu title={menuHeader}>
-                {internalLinks?.map((link) => (
-                    <AccordionAnchorMenuItem href={`#${link.anchorId}`}>
-                        {link.linkText}
-                    </AccordionAnchorMenuItem>
-                ))}
-            </AccordionAnchorMenu>
+            <PageNavigationMenu
+                title={menuHeader}
+                anchorLinks={internalLinks}
+            />
             <Region pageProps={pageProps} regionProps={regionProps} />
         </div>
     );
