+++ src/components/parts/PartsMapper.tsx
@@ -27,6 +27,7 @@ import { BEM, classNames } from '../../utils/classnames';
 import { HtmlArea } from './_dynamic/html-area/HtmlArea';
 import { PageHeaderPart } from './_dynamic/page-header/PageHeaderPart';
 import { ButtonPart } from './_dynamic/button/ButtonPart';
+import { PageNavigationMenuPart } from './page-navigation-menu/PageNavigationMenuPart';
 
 type Props = {
     partProps: PartComponentProps;
@@ -61,6 +62,7 @@ const partsWithOwnData: {
     [PartType.HtmlArea]: HtmlArea,
     [PartType.PageHeader]: PageHeaderPart,
     [PartType.Button]: ButtonPart,
+    [PartType.PageNavigationMenu]: PageNavigationMenuPart,
 };
 
 const partsDeprecated: { [key in PartDeprecated] } = {
