+++ src/types/component-props/parts.ts
@@ -24,6 +24,7 @@ export enum PartType {
     HtmlArea = 'no.nav.navno:html-area',
     PageHeader = 'no.nav.navno:page-header',
     Button = 'no.nav.navno:button',
+    PageNavigationMenu = 'no.nav.navno:page-navigation-menu',
 }
 
 export type PartDeprecated =
@@ -53,4 +54,5 @@ export type PartWithOwnData =
     | PartType.NewsList
     | PartType.HtmlArea
     | PartType.PageHeader
-    | PartType.Button;
+    | PartType.Button
+    | PartType.PageNavigationMenu;
