+++ src/types/component-props/parts/page-navigation-menu.ts
@@ -1,0 +1,15 @@
+import { PartComponentProps } from '../_component-common';
+import { PartType } from '../parts';
+
+export type AnchorLink = {
+    anchorId: string;
+    linkText: string;
+};
+
+export interface PageNavigationMenuProps extends PartComponentProps {
+    descriptor: PartType.PageNavigationMenu;
+    config: {
+        title: string;
+        anchorLinks: AnchorLink[];
+    };
+}
